Выполняет предварительное разделение на модули
